pxar: separate LOG macros from logger
